+++ tests/hromap/test_hromap_insert.c
@@ -151,13 +151,13 @@ CHECK_BEGIN_STATIC_FN(hromap_test_handle_api_functional)
     /* Over allocate size now because we don't want to worry about resizing. */
     ccc_handle_realtime_ordered_map hrm
         = hrm_init((struct val[200]){}, elem, id, id_cmp, NULL, NULL, 200);
-    size_t const size = 200;
+    ptrdiff_t const size = 200;
 
     /* Test handle or insert with for all even values. Default should be
        inserted. All entries are hashed to last digit so many spread out
        collisions. */
     struct val def = {0};
-    for (size_t i = 0; i < size / 2; i += 2)
+    for (ptrdiff_t i = 0; i < size / 2; i += 2)
     {
         def.id = (int)i;
         def.val = (int)i;
@@ -169,7 +169,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_handle_api_functional)
     }
     CHECK(size(&hrm), (size / 2) / 2);
     /* The default insertion should not occur every other element. */
-    for (size_t i = 0; i < size / 2; ++i)
+    for (ptrdiff_t i = 0; i < size / 2; ++i)
     {
         def.id = (int)i;
         def.val = (int)i;
@@ -194,7 +194,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_handle_api_functional)
     CHECK(size(&hrm), (size / 2));
     /* More simply modifications don't require the and modify function. All
        should be switched back to even now. */
-    for (size_t i = 0; i < size / 2; ++i)
+    for (ptrdiff_t i = 0; i < size / 2; ++i)
     {
         def.id = (int)i;
         def.val = (int)i;
@@ -211,7 +211,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_handle_api_functional)
 CHECK_BEGIN_STATIC_FN(hromap_test_insert_via_handle)
 {
     /* Over allocate size now because we don't want to worry about resizing. */
-    size_t const size = 200;
+    ptrdiff_t const size = 200;
     ccc_handle_realtime_ordered_map hrm
         = hrm_init((struct val[200]){}, elem, id, id_cmp, NULL, NULL, 200);
 
@@ -219,7 +219,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_insert_via_handle)
        inserted. All entries are hashed to last digit so many spread out
        collisions. */
     struct val def = {0};
-    for (size_t i = 0; i < size / 2; i += 2)
+    for (ptrdiff_t i = 0; i < size / 2; i += 2)
     {
         def.id = (int)i;
         def.val = (int)i;
@@ -231,7 +231,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_insert_via_handle)
     }
     CHECK(size(&hrm), (size / 2) / 2);
     /* The default insertion should not occur every other element. */
-    for (size_t i = 0; i < size / 2; ++i)
+    for (ptrdiff_t i = 0; i < size / 2; ++i)
     {
         def.id = (int)i;
         def.val = (int)i + 1;
@@ -256,14 +256,14 @@ CHECK_BEGIN_STATIC_FN(hromap_test_insert_via_handle)
 CHECK_BEGIN_STATIC_FN(hromap_test_insert_via_handle_macros)
 {
     /* Over allocate size now because we don't want to worry about resizing. */
-    size_t const size = 200;
+    ptrdiff_t const size = 200;
     ccc_handle_realtime_ordered_map hrm
         = hrm_init((struct val[200]){}, elem, id, id_cmp, NULL, NULL, 200);
 
     /* Test handle or insert with for all even values. Default should be
        inserted. All entries are hashed to last digit so many spread out
        collisions. */
-    for (size_t i = 0; i < size / 2; i += 2)
+    for (ptrdiff_t i = 0; i < size / 2; i += 2)
     {
         struct val const *const d
             = hrm_at(&hrm, insert_handle(handle_r(&hrm, &i),
@@ -274,7 +274,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_insert_via_handle_macros)
     }
     CHECK(size(&hrm), (size / 2) / 2);
     /* The default insertion should not occur every other element. */
-    for (size_t i = 0; i < size / 2; ++i)
+    for (ptrdiff_t i = 0; i < size / 2; ++i)
     {
         struct val const *const d
             = hrm_at(&hrm, insert_handle(handle_r(&hrm, &i),
@@ -359,7 +359,8 @@ CHECK_BEGIN_STATIC_FN(hromap_test_two_sum)
     int const addends[10] = {1, 3, -980, 6, 7, 13, 44, 32, 995, -1};
     int const target = 15;
     int solution_indices[2] = {-1, -1};
-    for (size_t i = 0; i < (sizeof(addends) / sizeof(addends[0])); ++i)
+    for (ptrdiff_t i = 0; i < (ptrdiff_t)(sizeof(addends) / sizeof(addends[0]));
+         ++i)
     {
         struct val const *const other_addend
             = hrm_at(&hrm, get_key_val(&hrm, &(int){target - addends[i]}));
@@ -380,7 +381,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_two_sum)
 
 CHECK_BEGIN_STATIC_FN(hromap_test_resize)
 {
-    size_t const prime_start = 11;
+    ptrdiff_t const prime_start = 11;
     ccc_handle_realtime_ordered_map hrm
         = hrm_init((struct val *)malloc(sizeof(struct val) * prime_start), elem,
                    id, id_cmp, std_alloc, NULL, prime_start);
@@ -416,7 +417,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_resize)
 
 CHECK_BEGIN_STATIC_FN(hromap_test_resize_macros)
 {
-    size_t const prime_start = 11;
+    ptrdiff_t const prime_start = 11;
     ccc_handle_realtime_ordered_map hrm
         = hrm_init((struct val *)malloc(sizeof(struct val) * prime_start), elem,
                    id, id_cmp, std_alloc, NULL, prime_start);
@@ -551,7 +552,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_insert_limit)
         CHECK(v->val, i);
         last_index = shuffled_index;
     }
-    size_t const final_size = size(&hrm);
+    ptrdiff_t const final_size = size(&hrm);
     /* The last successful handle is still in the table and is overwritten. */
     struct val v = {.id = last_index, .val = -1};
     ccc_handle hndl = swap_handle(&hrm, &v.elem);
@@ -629,7 +630,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_insert_and_find)
 
 CHECK_BEGIN_STATIC_FN(hromap_test_insert_shuffle)
 {
-    size_t const size = 50;
+    ptrdiff_t const size = 50;
     ccc_handle_realtime_ordered_map hrm
         = hrm_init((struct val[51]){}, elem, id, id_cmp, NULL, NULL, 51);
     CHECK(size > 1, true);
@@ -637,7 +638,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_insert_shuffle)
     CHECK(insert_shuffled(&hrm, size, prime), PASS);
     int sorted_check[50];
     CHECK(inorder_fill(sorted_check, size, &hrm), size);
-    for (size_t i = 1; i < size; ++i)
+    for (ptrdiff_t i = 1; i < size; ++i)
     {
         CHECK(sorted_check[i - 1] <= sorted_check[i], true);
     }
@@ -657,7 +658,7 @@ CHECK_BEGIN_STATIC_FN(hromap_test_insert_weak_srand)
         CHECK(insert_error(&e), false);
         CHECK(validate(&hrm), true);
     }
-    CHECK(size(&hrm), (size_t)num_nodes);
+    CHECK(size(&hrm), (ptrdiff_t)num_nodes);
     CHECK_END_FN();
 }
